requirements.txt
fastapi
uvicorn
python-dotenv
pytest
httpx
loguru
pymongo